Patched error when unavailable file

+++ Elogind/246.10.0/246.10.0.cr
@@ -40,8 +40,11 @@ class Target < ISM::Software
 
         makeDirectory("#{builtSoftwareDirectoryPath}#{Ism.settings.rootPath}etc/pam.d")
 
-
-        copyFile("#{Ism.settings.rootPath}etc/pam.d/system-session","#{builtSoftwareDirectoryPath(false)}#{Ism.settings.rootPath}etc/pam.d/system-session")
+        if File.exists?("#{Ism.settings.rootPath}etc/pam.d/system-session")
+            copyFile("#{Ism.settings.rootPath}etc/pam.d/system-session","#{builtSoftwareDirectoryPath(false)}#{Ism.settings.rootPath}etc/pam.d/system-session")
+        else
+            generateEmptyFile("#{builtSoftwareDirectoryPath(false)}#{Ism.settings.rootPath}etc/pam.d/system-session")
+        end
 
         systemSessionData = <<-CODE
         session  required    pam_loginuid.so
